Fxxes CI and tests
- fixes tests that slipped in with recent commits
- drops support for Python 3.2 and 3.3
- adds support for Python 3.5
- run tests with pytest
